pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vigil"
version = "0.1.0"
description = "Behavioural verification of evolving systems under controlled variation."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
  { name = "Leopold Lemmermann" }
]
dependencies = [
  "PyYAML>=6.0",
  "python-dotenv>=1.0",
]

[project.scripts]
vigil = "core.cli:start"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
include = ["core*", "builtin*"]

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
]

[tool.pytest.ini_options]
testpaths = ["test"]